финальные правки: faker, переименования, allure, разделение тестов
